05_Content_Factory_V2/requirements.txt
elevenlabs>=1.0.0
moviepy==1.0.3
playwright>=1.40.0
Pillow>=10.0.0
requests
